[feat]: event code refactor
