+++ apps/client/app/page.tsx
@@ -1,12 +1,12 @@
 "use client"
 import HomeWrapper from "@/components/landing/Home/HomeWrapper";
-// import OrderBook from "@/components/landing/Orderbook/Orderbook";
+// import TradePage from "@/components/landing/EventTrade/Trade";
 
 export default function Page() {
   return (
     <div>
       <HomeWrapper/>
-      {/* <OrderBook eventId="jbdjfdf"/> */}
+      {/* <TradePage eventId="jbdjfdf"/> */}
     </div>
   );
 }
